chore: reorganized files for edge detection functions
