+++ models/modules/sketch_generation/mbv2_mlsd_large.py
